+++ R/SciViews-package.R
@@ -28,15 +28,15 @@
 
 #' @import stats
 #' @import graphics
+#' @importFrom utils packageVersion stack
 #' @importFrom grDevices colorRampPalette chull hsv rainbow
 #' @importFrom ellipse ellipse plotcorr
 #' @importFrom crayon num_colors num_ansi_colors white black red blue green bold col_align col_nchar
-#' @importFrom cli rule symbol
-#' @import rstudioapi
+#' @importFrom cli cat_line rule symbol
+#' @importFrom rstudioapi getThemeInfo hasFun isAvailable
+#' @importFrom purrr compact imap keep map map_chr map2_chr set_names
 #'
-#' @import tidyverse
-#' @import svMisc
-#' @import svFlow
+#' @importFrom svBase as_dtf as_dtt as_dtbl as_dtx is_dtf is_dtt is_dtbl
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start